fix: trust nothing in the managed Windows config path that a user controls
The managed config directory comes from the ProgramData known folder,
not from PROGRAMDATA in the process environment, which a user can set to
point PMG away from the config that governs them. Before setup reads an
existing config.yml to merge it, the file must already be
administrator-only. A user-writable file stops the install and stays as
evidence. A link or junction anywhere in the PMG-owned path stops it too.

The ACL validators move from shim to fsutil, next to ForceRootOwned, so
config can use them. They now reject an entry of a type they do not
evaluate, and a deny of execution to any principal outside the
administrator set.
