+++ config/config.go
@@ -896,11 +896,7 @@ func globalConfigDir() string {
 	case "linux":
 		return "/etc/safedep/pmg"
 	case "windows":
-		programData := os.Getenv("PROGRAMDATA")
-		if programData == "" {
-			programData = `C:\ProgramData`
-		}
-		return filepath.Join(programData, "safedep", "pmg")
+		return filepath.Join(programDataDir(), "safedep", "pmg")
 	}
 
 	return ""
@@ -1170,14 +1166,28 @@ func WriteSystemTemplateConfig() error {
 	// them. Only directories created here and the file we write are touched;
 	// pre-existing directories keep their permissions.
 	dir := filepath.Dir(path)
+
+	// ProgramData lets a standard user create a directory, and one created
+	// that way stays theirs. On Windows both PMG-owned components are
+	// therefore forced administrator-only even when they already exist, and
+	// nothing in the path may be a link, or the writes below would land
+	// where that user pointed them. An existing config file is merged only
+	// when it is already administrator-only: an owner set afterwards would
+	// not make contents a standard user wrote trustworthy. /etc has no such
+	// hole, so on Linux these checks are no-ops and pre-existing directories
+	// keep their permissions.
+	for _, d := range []string{filepath.Dir(dir), dir} {
+		if err := fsutil.RequireNotReparsePoint(d); err != nil {
+			return err
+		}
+	}
+	if err := fsutil.RequireTrustedExisting(path); err != nil {
+		return fmt.Errorf("%w. Inspect the file, delete it, and run the install again", err)
+	}
+
 	if err := fsutil.MkdirAllRootOwned(dir, 0o755); err != nil {
 		return err
 	}
-
-	// ProgramData lets a standard user create a directory, and one created
-	// that way stays theirs, so on Windows both PMG-owned components are
-	// forced administrator-only even when they already exist. /etc has no
-	// such hole, so Linux keeps pre-existing directories as they are.
 	if runtime.GOOS == "windows" {
 		for _, d := range []string{filepath.Dir(dir), dir} {
 			if err := fsutil.ForceRootOwned(d, 0o755); err != nil {
