+++ config/paths_other.go
@@ -1,0 +1,6 @@
+//go:build !windows
+
+package config
+
+// programDataDir is a Windows location.
+func programDataDir() string { return "" }
